Markers and visualization (#622)
* Marker optimization ; DB save/load ; init from multiple observations

Loading/saving marker info, 2D and 3D, to sqlite3

keep_fixed_ flag in marker: previously, the marker positions were not
optimized, the position was estimated once based on a single observation
and kept fixed. The default is now to include the position in the
optimization, which could still be turned off by setting the flag to
true (would currently need to be done in the code)

initialized_before_ flag in marker and
"required_keyframes_for_marker_initialization" setting: if a single
observation of the marker is used to estimate its 3D position, a bad
estimate can still throw off the optimization procedure, causing the
markers to mess up the map building instead of helping. The setting
allows the initialization to wait until a number of keyframes have seen
the marker. Only when this happens, the flag is set, and the marker will
be included in the optimization. The default is to wait for 3 frames,
but can be controlled using a config file setting.

* Aruco nano support

This patch add support for using the aruco nano header file
(https://sourceforge.net/projects/aruco/files/ArucoNano-v5/) to
detect markers. The user can enable the support using a CMake
config setting, and then use "aruconano" as marker model.

* Modifications for marker visualisation

This draws the 2D detected markers, together with their IDs,
on the observed image. Because this needs the original coordinates
(not the undistorted ones), they are stored in the marker2d
class now as well (not saved/loaded to database though). To
allow a publisher (eg the socket_publisher) to share the 3D marker
info, a method has been added to retrieve them.

* Clearing markers as well in 'clear()'

* Refactoring and some minor fixes

- Delete references from markers when erasing keyframes
- required_keyframes_for_marker_initialization can now be set individusually in KeyframeInseter and Initialiser

---------

Co-authored-by: Jori Liesenborgs <jori.liesenborgs@gmail.com>
Co-authored-by: ymd-stella <world.applepie@gmail.com>